+++ package.json
@@ -1,18 +1,20 @@
 {
-  "name": "TaskManager",
+  "name": "task-manager",
   "version": "1.0.0",
   "private": true,
   "description": "Личный проект «Менеджер задач» от HTML Academy",
   "scripts": {
     "build": "webpack",
     "eslint": "eslint src/**/*.js",
-    "test": "npm run eslint"
+    "test": "npm run eslint",
+    "start": "webpack-dev-server --open"
   },
   "devDependencies": {
     "eslint": "5.13.0",
     "eslint-config-htmlacademy": "0.5.2",
     "webpack": "4.29.5",
-    "webpack-cli": "3.2.3"
+    "webpack-cli": "3.2.3",
+    "webpack-dev-server": "3.2.1"
   },
   "engines": {
     "node": "10"
